requirements/requirements.txt
Flask==3.0.0
authlib==1.3.0
requests==2.31.0
PyJWT==2.8.0
python-dotenv==1.0.0
cryptography==41.0.7
